Калькулятор - BJS2-5255, исправил недочеты задание отправляю PR

+++ src/main/java/Calculator/Calculator.java
@@ -2,5 +2,5 @@ package Calculator;
 
 @FunctionalInterface
 public interface Calculator {
-    int calculate(int a, int b);
+    long calculate(long a, long b);
 }
